requirements.txt
requests
pandas
python-dotenv
beautifulsoup4
lxml
woocommerce
